feature(mailchimp): moving functions to utils.js
