+++ __tests__/data/mailchimp_output.json
@@ -16,8 +16,152 @@
           "LNAME": "Dole",
           "anonymousId": "userId12345"
         },
+        "status": "subscribed",
         "userId": "userId12345",
         "email_address": "bob.dole@initech.com"
+      },
+      "XML": {},
+      "JSON_ARRAY": {},
+      "FORM": {}
+    },
+    "files": {},
+    "userId": "userId12345"
+  },
+  {
+    "version": "1",
+    "type": "REST",
+    "method": "PUT",
+    "endpoint": "https://us20.api.mailchimp.com/3.0/lists/df42a82d07/members/48cd6232dc124497369f59c33d3eb4ab",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": "Basic YXBpS2V5Ojk0ZjcxOTE3ZDg1MjI3NzBjOTc0NDliMGM5MGNhYTRjLXVzMjA="
+    },
+    "params": {},
+    "body": {
+      "JSON": {
+        "status": "subscribed",
+        "userId": "userId12345",
+        "email_address": "bob.dole@initech.com",
+        "merge_fields": {
+          "anonymousId": "userId12345"
+        }
+      },
+      "XML": {},
+      "JSON_ARRAY": {},
+      "FORM": {}
+    },
+    "files": {},
+    "userId": "userId12345"
+  },
+  {
+    "error": "User does not have access to the requested operation"
+  },
+  {
+    "error": "Message type page is not supported"
+  },
+  {
+    "version": "1",
+    "type": "REST",
+    "method": "PUT",
+    "endpoint": "https://us20.api.mailchimp.com/3.0/lists/df42a82d07/members/48cd6232dc124497369f59c33d3eb4ab",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": "Basic YXBpS2V5Ojk0ZjcxOTE3ZDg1MjI3NzBjOTc0NDliMGM5MGNhYTRjLXVzMjA="
+    },
+    "params": {},
+    "body": {
+      "JSON": {
+        "merge_fields": {
+          "FIRSTNAME": "Bob"
+        },
+        "status": "subscribed",
+        "email_address": "bob.dole@initech.com"
+      },
+      "XML": {},
+      "JSON_ARRAY": {},
+      "FORM": {}
+    },
+    "files": {},
+    "userId": "userId12345"
+  },
+  {
+    "error": "The status must be one of [subscribed, unsubscribed, cleaned, pending, transactional]"
+  },
+  {
+    "statusCode": 400,
+    "error": "email is required for identify",
+    "statTags": {
+      "destination": "mailchimp",
+      "stage": "transform",
+      "scope": "exception"
+    }
+  },
+  {
+    "body": {
+      "XML": {},
+      "FORM": {},
+      "JSON": {
+        "status": "subscribed",
+        "userId": "userId12345",
+        "merge_fields": {},
+        "email_address": "jhon@gmail.com"
+      },
+      "JSON_ARRAY": {}
+    },
+    "type": "REST",
+    "files": {},
+    "method": "POST",
+    "params": {},
+    "userId": "userId12345",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": "Basic YXBpS2V5Ojk0ZjcxOTE3ZDg1MjI3NzBjOTc0NDliMGM5MGNhYTRjLXVzMjA="
+    },
+    "version": "1",
+    "endpoint": "https://us20.api.mailchimp.com/3.0/lists/ff21810eec/members"
+  },
+  {
+    "body": {
+      "XML": {},
+      "FORM": {},
+      "JSON": {
+        "merge_fields": {},
+        "userId": "userId12345",
+        "email_address": "bob.dole@initech.com",
+        "status": "subscribed"
+      },
+      "JSON_ARRAY": {}
+    },
+    "type": "REST",
+    "files": {},
+    "method": "PUT",
+    "params": {},
+    "userId": "userId12345",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": "Basic YXBpS2V5Ojk0ZjcxOTE3ZDg1MjI3NzBjOTc0NDliMGM5MGNhYTRjLXVzMjA="
+    },
+    "version": "1",
+    "endpoint": "https://us20.api.mailchimp.com/3.0/lists/df42a82d07/members/48cd6232dc124497369f59c33d3eb4ab"
+  },
+  {
+    "version": "1",
+    "type": "REST",
+    "method": "PUT",
+    "endpoint": "https://us20.api.mailchimp.com/3.0/lists/df42a82d07/members/48cd6232dc124497369f59c33d3eb4ab",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": "Basic YXBpS2V5Ojk0ZjcxOTE3ZDg1MjI3NzBjOTc0NDliMGM5MGNhYTRjLXVzMjA="
+    },
+    "params": {},
+    "body": {
+      "JSON": {
+        "status": "subscribed",
+        "userId": "userId12345",
+        "email_address": "bob.dole@initech.com",
+        "merge_fields": {
+          "anonymousId": "userId12345"
+        }
       },
       "XML": {},
       "JSON_ARRAY": {},
